backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sqlalchemy
python-multipart
pydantic
python-dotenv
requests
plotly
streamlit